fixing bugs with report

+++ app/models/report.rb
@@ -11,7 +11,7 @@ class Report < ActiveRecord::Base
   def self.plates(tipo, fecha, estados)
     #FileUtils.rm_rf("public/pdf/")   
     Dir.mkdir("public/pdf/") unless File.exists?("public/pdf/")
-    now = Time.now
+    now = Time.zone.now
     @name = now.to_s + " " + getHorario(tipo) + ".pdf"
     order_count = 0   
     Prawn::Document.generate("public/pdf/"+ @name, :page_layout => :landscape ) do |pdf|            
@@ -54,16 +54,17 @@ class Report < ActiveRecord::Base
 
   end
 
-  def self.areas(tipo, fecha)
-    FileUtils.rm_rf("public/pdf/")   
+  def self.areas(tipo, fecha)    
     Dir.mkdir("public/pdf/") unless File.exists?("public/pdf/")
-    now = Time.now
-    @name = now.to_s + " Pedidos Area.pdf"    
+    now = Time.zone.now
+    @name = now.to_s + " Pedidos Area.pdf"
+    count_platos = 0
     Prawn::Document.generate("public/pdf/"+ @name ) do         
       Area.all.each do |area|
         platos_info = Array.new
         Area.get_plates_by_horario_and_area(tipo.to_i, area.id, fecha).each do |jn|
-          platos_info << [jn.regimen, jn.plato, jn.numero.to_s]
+          platos_info << [jn.regimen, jn.plato, jn.numero.to_s] 
+          count_platos = count_platos + 1         
         end       
         if platos_info.count > 0
           text_box "PEDIDOS POR AREA", :at => [85, 720], :width => 300, :align => :center, :size => 20
@@ -127,6 +128,7 @@ class Report < ActiveRecord::Base
               transparent(0.8) { stroke_line [380, 615], [380, 0] }               
               z = 560
               cambios.each do |cmb|
+                count_platos = count_platos + 1
                 nombre = Plate.find(cmb.plate_id).nombre.split("[")[0]
                 text_box "#{nombre}", :at => [0, z], :width => 375, :align => :left, :size => 12
                 if cmb.number.to_i < 0
@@ -148,6 +150,8 @@ class Report < ActiveRecord::Base
           start_new_page
         end
       end
+      #si no tiene nada retorna
+      return "0" if count_platos == 0
       #borrar los estado areas y logs especificos
       EstadoArea.find_all_by_fecha_and_horario(fecha, tipo.to_i).each do |ear|
         ear.destroy
@@ -233,7 +237,7 @@ class Report < ActiveRecord::Base
 
   def self.colacion(tipo, fecha, estados)
     Dir.mkdir("public/pdf/") unless File.exists?("public/pdf/")
-    now = Time.now
+    now = Time.zone.now
     @name = now.to_s + " " + "Colacion" + ".pdf"
     order_count = 0
     @orders = OrderList.joins(:orders).where(:fecha => fecha, :orders => { :horario => tipo, :estado => estados }).select("order_lists.patient_id, orders.horario, orders.comentarios, orders.id as order_id")
